requirements.txt
gitdb==4.0.7
GitPython==3.1.18
PyGithub==1.55
requests==2.25.1
tabulate==0.8.9
